allow for Application/src directory autoloading
Unless I am missing something, it does not appear that the Application/src folder is autoloaded. This would allow for the following:

$test = new Application\Core\Folder\File();

which would autoload the File class from

application/src/Folder/File.php

Anything outside of the Application\Core namespace would be left alone. Again, I'm not sure if I wasn't namespacing, or loading the classes correctly if this functionality already exists.